trading-web-app/backend/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
ib_insync==0.9.84
python-dotenv==1.0.0
